requirements.txt
blinker==1.8.2
capstone==5.0.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.7.0
Flask==3.0.3
future==1.0.0
idna==3.10
impacket==0.12.0
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pyOpenSSL==24.0.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.4
pytz==2024.2
requests==2.32.3
scapy==2.5.0
setuptools==75.1.0
six==1.16.0
Sublist3r==1.0
thunderclient==1.0.1
urllib3==2.2.3
Werkzeug==3.0.4
